+++ Modules/AOC2019.Modules.Intcode/IntcodeOutput.cs
@@ -4,21 +4,21 @@ using System.Linq;
 
 namespace AOC2019.Modules.Intcode
 {
-    public class IntcodeOutput : IEnumerable<int>
+    public class IntcodeOutput : IEnumerable<long>
     {
-        private readonly List<int> _outputs;
+        private readonly List<long> _outputs;
 
-        public IntcodeOutput(IEnumerable<int> outputs)
+        public IntcodeOutput(IEnumerable<long> outputs)
         {
             _outputs = outputs.ToList();
         }
 
-        public int this[int index] => _outputs[index];
-        public void Add(int val) => _outputs.Add(val);
+        public long this[int index] => _outputs[index];
+        public void Add(long val) => _outputs.Add(val);
 
 
         public override string ToString() => string.Join(",", _outputs);
-        public IEnumerator<int> GetEnumerator() => ((IEnumerable<int>)_outputs).GetEnumerator();
-        IEnumerator IEnumerable.GetEnumerator() => ((IEnumerable<int>)_outputs).GetEnumerator();
+        public IEnumerator<long> GetEnumerator() => ((IEnumerable<long>)_outputs).GetEnumerator();
+        IEnumerator IEnumerable.GetEnumerator() => ((IEnumerable<long>)_outputs).GetEnumerator();
     }
 }
